Back to ArrayList:(
HashMaps do not work. Detection does now though!

+++ Obj.java
@@ -5,7 +5,7 @@ public class Obj {
 
     public Obj() {
         color = Color.WHITE;
-        position = new Coor(0.0, 0.0);
+        position = new Coor(0, 0);
     }
 
     public Obj(Color color, Coor position) {
@@ -14,11 +14,11 @@ public class Obj {
     }
 
     // setters 
-    public void setPosX(double x) {
+    public void setPosX(int x) {
         position.setX(x);
     }
 
-    public void setPosY(double y) {
+    public void setPosY(int y) {
         position.setY(y);
     }
 
@@ -31,11 +31,11 @@ public class Obj {
         return position;
     }
 
-    public double getPosY() {
+    public int getPosY() {
         return position.y(); 
     }
 
-    public double getPosX() {
+    public int getPosX() {
         return position.x(); 
     }
     
@@ -46,7 +46,7 @@ public class Obj {
     public Coor getPrintPos() {
         Translation t = new Translation();
         t.setMat(Main.frame.getWidth()/1000.0, Main.frame.getHeight()/1000.0);
-        double[] ans = t.translate(this.getPos().matrix());
+        int[] ans = t.translate(this.getPos().matrix());
         return new Coor(ans[0], ans[1]);
     }
